+++ vortex-fastlanes/benches/bitpacking.rs
@@ -1,19 +1,9 @@
-use std::sync::Arc;
-
-use criterion::{black_box, criterion_group, criterion_main, Criterion};
-use fastlanez::TryBitPack;
-use itertools::Itertools;
+use criterion::{black_box, Criterion, criterion_group, criterion_main};
+use rand::{Rng, thread_rng};
 use rand::distributions::Uniform;
-use rand::{thread_rng, Rng};
-use vortex::array::downcast::DowncastArrayBuiltin;
-use vortex::array::primitive::PrimitiveArray;
-use vortex::compress::{CompressConfig, CompressCtx, EncodingCompression};
-use vortex::compute::take::take;
-use vortex::encoding::EncodingRef;
-use vortex_fastlanes::{
-    bitpack_primitive, unpack_primitive, unpack_single_primitive, BitPackedEncoding,
-    DowncastFastlanes,
-};
+
+use fastlanez::TryBitPack;
+use vortex_fastlanes::{bitpack_primitive, unpack_primitive, unpack_single_primitive};
 
 fn values(len: usize, bits: usize) -> Vec<u32> {
     let rng = thread_rng();
@@ -71,77 +61,5 @@ fn pack_unpack(c: &mut Criterion) {
     });
 }
 
-fn bench_take(c: &mut Criterion) {
-    let cfg = CompressConfig::new().with_enabled([&BitPackedEncoding as EncodingRef]);
-    let ctx = CompressCtx::new(Arc::new(cfg));
-
-    let values = values(1_000_000, 8);
-    let uncompressed = PrimitiveArray::from(values.clone());
-    let packed = BitPackedEncoding {}
-        .compress(&uncompressed, None, ctx)
-        .unwrap();
-
-    let stratified_indices: PrimitiveArray = (0..10).map(|i| i * 10_000).collect::<Vec<_>>().into();
-    c.bench_function("take_10_stratified", |b| {
-        b.iter(|| black_box(take(&packed, &stratified_indices).unwrap()));
-    });
-
-    let contiguous_indices: PrimitiveArray = (0..10).collect::<Vec<_>>().into();
-    c.bench_function("take_10_contiguous", |b| {
-        b.iter(|| black_box(take(&packed, &contiguous_indices).unwrap()));
-    });
-
-    let rng = thread_rng();
-    let range = Uniform::new(0, values.len());
-    let random_indices: PrimitiveArray = rng
-        .sample_iter(range)
-        .take(10_000)
-        .map(|i| i as u32)
-        .collect_vec()
-        .into();
-    c.bench_function("take_10K_random", |b| {
-        b.iter(|| black_box(take(&packed, &random_indices).unwrap()));
-    });
-
-    let contiguous_indices: PrimitiveArray = (0..10_000).collect::<Vec<_>>().into();
-    c.bench_function("take_10K_contiguous", |b| {
-        b.iter(|| black_box(take(&packed, &contiguous_indices).unwrap()));
-    });
-}
-
-fn bench_take_patched(c: &mut Criterion) {
-    let cfg = CompressConfig::new().with_enabled([&BitPackedEncoding as EncodingRef]);
-    let ctx = CompressCtx::new(Arc::new(cfg));
-
-    let big_base2 = 1048576;
-    let num_exceptions = 100;
-    let values = (0u32..big_base2 + num_exceptions).collect_vec();
-
-    let uncompressed = PrimitiveArray::from(values.clone());
-    let packed = BitPackedEncoding {}
-        .compress(&uncompressed, None, ctx)
-        .unwrap();
-    let packed = packed.as_bitpacked();
-    assert!(packed.patches().is_some());
-    assert_eq!(
-        packed.patches().unwrap().as_sparse().values().len(),
-        num_exceptions as usize
-    );
-
-    let not_patch_indices: PrimitiveArray = (0u32..10000).collect_vec().into();
-    c.bench_function("take_10K_not_patches", |b| {
-        b.iter(|| black_box(take(packed, &not_patch_indices).unwrap()));
-    });
-
-    let patch_indices: PrimitiveArray = (big_base2..big_base2 + num_exceptions)
-        .cycle()
-        .take(10000)
-        .collect_vec()
-        .into();
-    c.bench_function("take_10K_patches", |b| {
-        b.iter(|| black_box(take(packed, &patch_indices).unwrap()));
-    });
-}
-
-criterion_group!(benches, pack_unpack, bench_take, bench_take_patched);
+criterion_group!(benches, pack_unpack);
 criterion_main!(benches);
